Initial commit.  Add basic driver program.
This changeset introduces a Gemfile with the following
dependencies:

- elasticsearch
- sinatra
- pi_piper

The demo script ``driver.rb`` included attempts to expose
a few basic HTTP endpoints via Sinatra with the goal of
controlling a Raspberry Pi remotely.  Currently, the script
is tested using a single LED and resistor wired off of pin 2
on the B+ model.   See README for basic setup instructions.