Add configuration page with UI templating
This adds the Sinatra default /public and /views
directories, allowing us to serve up static content
like CSS and JS as well as render variables into our
view templates using eRubis (built-in).  This addition
goes on to migrate the root URL to a JSON health status
default response and to add a POST endpoint for
receiving the configuration data sent by the end-user.